requirements.txt
astroid==1.6.1
certifi==2018.1.18
chardet==3.0.4
cloudinary==1.11.0
colorama==0.3.9
Django==2.0.4
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-oauth-toolkit==1.0.0
django-redis==4.9.0
djangorestframework==3.8.2
googlemaps==3.0.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
mock==2.0.0
oauthlib==2.0.7
pbr==4.0.2
Pillow==5.1.0
pylint==1.8.2
pytz==2018.4
redis==2.10.6
requests==2.18.4
rope==0.11.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.11
